requirements.txt
aniso8601==9.0.1
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
Flask==2.0.3
Flask-RESTful==0.3.9
idna==3.3
iniconfig==1.1.1
itsdangerous==2.1.1
Jinja2==3.0.3
MarkupSafe==2.1.1
marshmallow==3.15.0
packaging==21.3
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.7
pytest==7.1.0
python-dotenv==0.19.2
pytz==2021.3
requests==2.27.1
six==1.16.0
tomli==2.0.1
urllib3==1.26.9
Werkzeug==2.0.3
